исправлены тесты (в новом тестовом проекте), старый будет удален
