+++ Theater_mdk.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.14.36414.22
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Theater_mdk", "Theater_mdk\Theater_mdk.csproj", "{A3E61A4A-E23B-4FA1-A57E-1CF12F406620}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Theater.Test", "Theater.Test\Theater.Test.csproj", "{AA5B2889-859D-407D-B38F-7B0E0C1EF66A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Theater_mdk.Test", "Theater_mdk.Test\Theater_mdk.Test.csproj", "{5D27D2C0-AB62-416C-A6BD-2BDD0BE5E4BF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{A3E61A4A-E23B-4FA1-A57E-1CF12F406620}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A3E61A4A-E23B-4FA1-A57E-1CF12F406620}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A3E61A4A-E23B-4FA1-A57E-1CF12F406620}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AA5B2889-859D-407D-B38F-7B0E0C1EF66A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AA5B2889-859D-407D-B38F-7B0E0C1EF66A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AA5B2889-859D-407D-B38F-7B0E0C1EF66A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AA5B2889-859D-407D-B38F-7B0E0C1EF66A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5D27D2C0-AB62-416C-A6BD-2BDD0BE5E4BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5D27D2C0-AB62-416C-A6BD-2BDD0BE5E4BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5D27D2C0-AB62-416C-A6BD-2BDD0BE5E4BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5D27D2C0-AB62-416C-A6BD-2BDD0BE5E4BF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
